+++ tests/e2e/initialization/config.go
@@ -481,13 +481,14 @@ func updateTWAPGenesis(appGenState map[string]json.RawMessage) func(twapGenState
 				}
 
 				twapRecord := twaptypes.TwapRecord{
-					PoolId:                      balancerPool.Id,
-					Asset0Denom:                 denomPair.Denom0,
-					Asset1Denom:                 denomPair.Denom0,
-					Height:                      1,
-					Time:                        time.Date(2023, 0o2, 1, 0, 0, 0, 0, time.UTC), // some time in the past.
-					P0LastSpotPrice:             sp0,
-					P1LastSpotPrice:             sp1,
+					PoolId:      balancerPool.Id,
+					Asset0Denom: denomPair.Denom0,
+					Asset1Denom: denomPair.Denom0,
+					Height:      1,
+					Time:        time.Date(2023, 0o2, 1, 0, 0, 0, 0, time.UTC), // some time in the past.
+					// Note: truncation is acceptable as x/twap is guaranteed to work only on pools with spot prices > 10^-18.
+					P0LastSpotPrice:             sp0.Dec(),
+					P1LastSpotPrice:             sp1.Dec(),
 					P0ArithmeticTwapAccumulator: osmomath.ZeroDec(),
 					P1ArithmeticTwapAccumulator: osmomath.ZeroDec(),
 					GeometricTwapAccumulator:    osmomath.ZeroDec(),